Change the `run` recipe to reflect intended use
The purpose of the default `run` Justfile recipe is to demonstrate
the general, anticipated and advocated usage of the program, i.e.
running the program with no arguments and piping its output into
a pager which will handle formatting.  We do not expect users to
generally need to pass arguments to the program every invocation.

Therefore this patch simplifies the `run` recipe, the default if we
invoke `just` with no recipe name, so that the `run` recipe accepts
no arguments and instead simply runs the program through a pager.
If we need to run the program with specific arguments for testing
purposes or some such then we should and will create recipes specific
to those situations.