Add the `--tip-file` option
The program requires a YAML document containing the tips it reads
from and displays to the user.  By default the environment variable
`TOTD_TIP_FILE` contains the absolute path to that document.  However,
the user may want to specify a different YAML document for a variety
of reasons.

Firstly, the user may have no easy way to define the `TOTD_TIP_FILE`
variable in his or her environment.  For example, the user may not
have permission to set such a variable or may not have access to a
program like `env` to define the variable.  We do not predict this
will a common limitation.  That is, we expect that the overwhelming
majority of users *will* be able to define this environment variable.
But in the rare situations where they cannot, they can use `--tip-file`
to name their preferred YAML document of tips.

Secondly, the user may wish to use a different YAML document without
(re)defining the `TOTD_TIP_FILE` environment variable.  For example,
the user may wish to temporarily use a different file that has tips
for specific software.  And developers of this program itself may
want to use a different file for testing and debugging purposes.
The new `--tip-file` option provides a straight-forward solution to
those problematic scenarios.

(Admittedly, this principle motivation for this change is the desire
to specify a separate document of tips for testing and debugging.)

A crucial change in this patch is that there is no longer implicit
default tip file.  Previously the program would try to use the file
`Tips.yaml`, relative to the program location, if there was no
`TOTD_TIP_FILE` environment variable.  This is no longer the case,
and the user now **must** either define that variable or use the
`--tip-file` option, as we no longer provide a default path to any
YAML file of tips.

Finally, this patch is the first step towards an intended feature
for a future version: support for multiple tip files.  In the first
paragraph of this commit message we mention the `env` program, which
we can use interchangably with our `--tip-file` option.  For example,
these two invocations of the program have exactly the same behavior:

    # Using the environment variable...
    $ env TOTD_TIP_FILE=/tmp/tips.yaml python3 totd.py

    # Using the new option...
    $ python3 totd.py --tip-file /tmp/tips.yaml

The "inconvenience" of using `env` is, by itself, not a strong enough
justification for introducing the `--tip-file` option.  However,
consider this hypothetical invocation:

    # Remember, this does not work.  Yet...
    python3 totd.py                                      \
        --tip-file $HOME/Projects/ToTD/Tips.yaml         \
        --tip-file $HOME/Documents/Tips/{Git,Emacs}.yaml \
        --tip-file /tmp/debugging-tips.yaml

There is no obviously intuitive way to achieve the same results using
`env` alone, or by using the `TOTD_TIP_FILE` variable.  Again, to be
clear, this patch does *not* implement support for multiple tip files
in any way.  But it does provide the new option with an eye towards
eventually supporting the hypothetical example given above.