Lint everything during `just push`
Our Justfile has a `push` recipe responsible for pushing updates to
the official GitHub repository.  This patch makes the `push` recipe
dependent on the `lint` recipe.  As a result `just push` will fail
to actually push anything if any of our lint-related recipes fail.
This helps protect us from pushing out a broken release, and while
not bullet-proof, it is better than nothing.