Allow multiple uses of the `--tip-file` option
Instead of having one monolithic YAML document of tips, users may wish
to organize their tips across multiple documents.  For example, they
may want one YAML document for Git tips, another for web browser tips,
and a third for general shell scripting tips.  Before this patch it
is not possible to do this, short of manually concatenating all of the
YAML documents prior to running ToTD.  This is an annoying limitation.

This patch is a large step towards removing that limitation by allowing
`--tip-file` to appear multiple times in a single invocation.  See the
new Justfile recipe for an example.  Instead of actually concatenating
the documents, which can create invalid YAML, the program will randomly
select a file of tips, and from *that* select a random tip.

(Note: The new `test-multiple-tip-files` Justfile recipe uses three
documents which are not included in the repository, and in fact we
tell Git to explicitly ignore them.  Eventually we may add these
documents depending on how we expand our unit testing.)

One remaining limitation, however, is that there is no way to make
ToTD select from multiple files solely through the `TOTD_TIP_FILE`
environment variable.  That is, the variable cannot represent
multiple files.  We plan to fix this in a future patch.